fix empty bag state render
